requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
daphne==3.0.1
Django==3.1
django-filter==2.4.0
djangochannelsrestframework==0.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
et-xmlfile==1.0.1
fpdf==1.7.2
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
msgpack==1.0.0
openpyxl==3.0.3
packaging==20.7
Pillow==7.2.0
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.0
pyparsing==2.4.7
pytest==6.1.2
pytest-asyncio==0.14.0
pytest-django==4.1.0
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.26.2
xlrd==1.2.0
zope.interface==5.2.0
